agent-foundry-api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-multipart>=0.0.6
celery>=5.3.0
pyyaml>=6.0.1
stripe>=8.0.0
litellm>=1.30.0
cryptography>=42.0.0
docker>=7.0.0
authlib>=1.3.0
httpx>=0.27.0
